# apps/web/src/content/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ClearGateway API",
    "version": "1.0.0",
    "description": "Non-custodial stablecoin payments (testnet)."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal API: ADMIN_TOKEN, plus an X-Admin-Email header naming a staff account."
      }
    },
...
  },
  "paths": {
    "/v1/payment_intents": {
      "post": {
        "summary": "Create a payment intent",
        "tags": [
          "payment_intents"
        ],
        "description": "Requires an `Idempotency-Key` header. `amount` is an integer string in USDC base units (6 decimals). The response includes `checkout_url` for the hosted checkout page.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{0,17}$"
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "USDC"
                    ],
                    "default": "USDC"
                  },
                  "merchant_order_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "maxLength": 500
                    },
                    "default": {}
                  },
                  "customer_email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 320
                  },
                  "success_url": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 2048
                  },
                  "cancel_url": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 2048
                  },
                  "expires_in_seconds": {
                    "type": "integer",
                    "minimum": 300,
                    "maximum": 86400,
                    "default": 3600
                  }
                },
                "required": [
                  "amount"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "object": {
                      "type": "string",
                      "enum": [
                        "payment_intent"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "test",
                        "live"
                      ]
                    },
                    "amount": {
                      "type": "string"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "decimals": {
                      "type": "number"
                    },
                    "fee_bps": {
                      "type": "number"
                    },
                    "fee_amount": {
                      "type": "string"
                    },
                    "merchant_amount": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "created",
                        "awaiting_payment",
                        "confirming",
                        "succeeded",
                        "expired",
                        "underpaid",
                        "failed",
                        "canceled",
                        "refunded",
                        "partially_refunded"
                      ]
                    },
                    "payment_method": {
                      "type": "string",
                      "nullable": true
                    },
                    "merchant_order_id": {
                      "type": "string",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "customer_email": {
                      "type": "string",
                      "nullable": true
                    },
                    "payer_address": {
                      "type": "string",
                      "nullable": true
                    },
                    "success_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "cancel_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "tx_hash": {
                      "type": "string",
                      "nullable": true
                    },
                    "confirmations": {
                      "type": "number"
                    },
                    "requires_review": {
                      "type": "boolean"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "succeeded_at": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "client_secret": {
                      "type": "string",
                      "nullable": true
                    },
                    "checkout_url": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "mode",
                    "amount",
                    "currency",
                    "decimals",
                    "fee_bps",
                    "fee_amount",
                    "merchant_amount",
                    "status",
                    "payment_method",
                    "merchant_order_id",
                    "metadata",
                    "customer_email",
                    "payer_address",
                    "success_url",
                    "cancel_url",
                    "tx_hash",
                    "confirmations",
                    "requires_review",
                    "expires_at",
                    "succeeded_at",
                    "created_at",
                    "client_secret",
                    "checkout_url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List payment intents (cursor pagination, newest first)",
        "tags": [
          "payment_intents"
        ],
        "description": "Pass the last id of a page as `starting_after` to fetch the next page.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^pi_[0-9a-f]{32}$"
            },
            "in": "query",
            "name": "starting_after",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "created",
                "awaiting_payment",
                "confirming",
                "succeeded",
                "expired",
                "underpaid",
                "failed",
                "canceled",
                "refunded",
                "partially_refunded"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "created_gte",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "created_lte",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "object": {
                      "type": "string",
                      "enum": [
                        "list"
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "object": {
                            "type": "string",
                            "enum": [
                              "payment_intent"
                            ]
                          },
                          "mode": {
                            "type": "string",
                            "enum": [
                              "test",
                              "live"
                            ]
                          },
                          "amount": {
                            "type": "string"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "decimals": {
                            "type": "number"
                          },
                          "fee_bps": {
                            "type": "number"
                          },
                          "fee_amount": {
                            "type": "string"
                          },
                          "merchant_amount": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "created",
                              "awaiting_payment",
                              "confirming",
                              "succeeded",
                              "expired",
                              "underpaid",
                              "failed",
                              "canceled",
                              "refunded",
                              "partially_refunded"
                            ]
                          },
                          "payment_method": {
                            "type": "string",
                            "nullable": true
                          },
                          "merchant_order_id": {
                            "type": "string",
                            "nullable": true
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "customer_email": {
                            "type": "string",
                            "nullable": true
                          },
                          "payer_address": {
                            "type": "string",
                            "nullable": true
                          },
                          "success_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "cancel_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "tx_hash": {
                            "type": "string",
                            "nullable": true
                          },
                          "confirmations": {
                            "type": "number"
                          },
                          "requires_review": {
                            "type": "boolean"
                          },
                          "expires_at": {
                            "type": "string"
                          },
                          "succeeded_at": {
                            "type": "string",
                            "nullable": true
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "client_secret": {
                            "type": "string",
                            "nullable": true
                          },
                          "checkout_url": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "object",
                          "mode",
                          "amount",
                          "currency",
                          "decimals",
                          "fee_bps",
                          "fee_amount",
                          "merchant_amount",
                          "status",
                          "payment_method",
                          "merchant_order_id",
                          "metadata",
                          "customer_email",
                          "payer_address",
                          "success_url",
                          "cancel_url",
                          "tx_hash",
                          "confirmations",
                          "requires_review",
                          "expires_at",
                          "succeeded_at",
                          "created_at",
                          "client_secret",
                          "checkout_url"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "object",
                    "data",
                    "has_more",
                    "next_cursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment_intents/{id}": {
      "get": {
        "summary": "Retrieve a payment intent",
        "tags": [
          "payment_intents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^pi_[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "object": {
                      "type": "string",
                      "enum": [
                        "payment_intent"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "test",
                        "live"
                      ]
                    },
                    "amount": {
                      "type": "string"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "decimals": {
                      "type": "number"
                    },
                    "fee_bps": {
                      "type": "number"
                    },
                    "fee_amount": {
                      "type": "string"
                    },
                    "merchant_amount": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "created",
                        "awaiting_payment",
                        "confirming",
                        "succeeded",
                        "expired",
                        "underpaid",
                        "failed",
                        "canceled",
                        "refunded",
                        "partially_refunded"
                      ]
                    },
                    "payment_method": {
                      "type": "string",
                      "nullable": true
                    },
                    "merchant_order_id": {
                      "type": "string",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "customer_email": {
                      "type": "string",
                      "nullable": true
                    },
                    "payer_address": {
                      "type": "string",
                      "nullable": true
                    },
                    "success_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "cancel_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "tx_hash": {
                      "type": "string",
                      "nullable": true
                    },
                    "confirmations": {
                      "type": "number"
                    },
                    "requires_review": {
                      "type": "boolean"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "succeeded_at": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "client_secret": {
                      "type": "string",
                      "nullable": true
                    },
                    "checkout_url": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "mode",
                    "amount",
                    "currency",
                    "decimals",
                    "fee_bps",
                    "fee_amount",
                    "merchant_amount",
                    "status",
                    "payment_method",
                    "merchant_order_id",
                    "metadata",
                    "customer_email",
                    "payer_address",
                    "success_url",
                    "cancel_url",
                    "tx_hash",
                    "confirmations",
                    "requires_review",
                    "expires_at",
                    "succeeded_at",
                    "created_at",
                    "client_secret",
                    "checkout_url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment_intents/{id}/cancel": {
      "post": {
        "summary": "Cancel an unpaid payment intent",
        "tags": [
          "payment_intents"
        ],
        "description": "Allowed only while `created` or `awaiting_payment`. A wallet payload already issued stays valid on-chain until its signature expiry; a payment landing after cancellation is flagged for review, never dropped.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^pi_[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "object": {
                      "type": "string",
                      "enum": [
                        "payment_intent"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "test",
                        "live"
                      ]
                    },
                    "amount": {
                      "type": "string"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "decimals": {
                      "type": "number"
                    },
                    "fee_bps": {
                      "type": "number"
                    },
                    "fee_amount": {
                      "type": "string"
                    },
                    "merchant_amount": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "created",
                        "awaiting_payment",
                        "confirming",
                        "succeeded",
                        "expired",
                        "underpaid",
                        "failed",
                        "canceled",
                        "refunded",
                        "partially_refunded"
                      ]
                    },
                    "payment_method": {
                      "type": "string",
                      "nullable": true
                    },
                    "merchant_order_id": {
                      "type": "string",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "customer_email": {
                      "type": "string",
                      "nullable": true
                    },
                    "payer_address": {
                      "type": "string",
                      "nullable": true
                    },
                    "success_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "cancel_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "tx_hash": {
                      "type": "string",
                      "nullable": true
                    },
                    "confirmations": {
                      "type": "number"
                    },
                    "requires_review": {
                      "type": "boolean"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "succeeded_at": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "client_secret": {
                      "type": "string",
                      "nullable": true
                    },
                    "checkout_url": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "mode",
                    "amount",
                    "currency",
                    "decimals",
                    "fee_bps",
                    "fee_amount",
                    "merchant_amount",
                    "status",
                    "payment_method",
                    "merchant_order_id",
                    "metadata",
                    "customer_email",
                    "payer_address",
                    "success_url",
                    "cancel_url",
                    "tx_hash",
                    "confirmations",
                    "requires_review",
                    "expires_at",
                    "succeeded_at",
                    "created_at",
                    "client_secret",
                    "checkout_url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}